sonos revert next alarm binary sensor to sensor
