+++ homeassistant/components/sonos/strings.json
@@ -20,9 +20,6 @@
   },
   "entity": {
     "binary_sensor": {
-      "alarm_scheduled": {
-        "name": "Alarm scheduled"
-      },
       "microphone": {
         "name": "Microphone"
       }
@@ -56,6 +53,9 @@
     "sensor": {
       "audio_input_format": {
         "name": "Audio input format"
+      },
+      "next_alarm": {
+        "name": "Next alarm"
       },
       "power_source": {
         "name": "Power source",
