+++ tests/components/sonos/snapshots/test_diagnostics.ambr
@@ -18,7 +18,6 @@
           'Temperature': 'NORMAL',
         }),
         'enabled_entities': list([
-          'binary_sensor.zone_a_alarm_scheduled',
           'binary_sensor.zone_a_charging',
           'binary_sensor.zone_a_microphone',
           'media_player.zone_a',
@@ -31,6 +30,7 @@
           'number.zone_a_treble',
           'sensor.zone_a_audio_input_format',
           'sensor.zone_a_battery',
+          'sensor.zone_a_next_alarm',
           'switch.sonos_alarm_14',
           'switch.zone_a_crossfade',
           'switch.zone_a_loudness',
@@ -111,7 +111,6 @@
       'Temperature': 'NORMAL',
     }),
     'enabled_entities': list([
-      'binary_sensor.zone_a_alarm_scheduled',
       'binary_sensor.zone_a_charging',
       'binary_sensor.zone_a_microphone',
       'media_player.zone_a',
@@ -124,6 +123,7 @@
       'number.zone_a_treble',
       'sensor.zone_a_audio_input_format',
       'sensor.zone_a_battery',
+      'sensor.zone_a_next_alarm',
       'switch.sonos_alarm_14',
       'switch.zone_a_crossfade',
       'switch.zone_a_loudness',
